Added LobbyView
LobbyView has actionListener for the buttons. Each button has a action
command containing the player being challenged and the name of the game.
Panels are also added into a LinkedList (ArrayList? HashMap?) for so
that it's the code look nicer when changing view. Next step is getting
player names and game names for lobbyView, using dummy values right now
for testing.